pdf-extract/pyproject.toml
[project]
name = "pdf-extract-mcp"
version = "0.1.0"
description = "MCP server for PDF extraction and summarization"
requires-python = ">=3.10"
dependencies = [
    "mcp>=0.9.0",
    "pymupdf>=1.23.0",
    "sumy>=0.11.0",
    "nltk>=3.8.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"
